api: refactor api server plugins dir
