+++ packages/api-server/__tests__/plugins/stats.ts
@@ -1,4 +1,4 @@
-/// <reference path="../../src/plugins/stats/stats.ts" />
+/// <reference path="../../src/plugins/stats.ts" />
 
 import { bootstrapTestDatabase, getTestDatabase, releaseTestDatabase } from '../helpers/db';
 import { bootstrapApp, getTestApp, releaseApp } from '../helpers/app';
