+++ packages/api-server/src/core/cache/Cache.ts
@@ -2,7 +2,7 @@
 import { CacheProvider } from './provider/CacheProvider';
 import {DateTime} from 'luxon'
 import pino from 'pino';
-import {cacheHitCounter, cacheQueryTimer, measure} from '../../plugins/metrics/metrics';
+import {cacheHitCounter, cacheQueryTimer, measure} from '../../plugins/metrics';
 
 export const MAX_CACHE_TIME = DateTime.fromISO('2099-12-31T00:00:00')
