+++ packages/api-server/src/core/runner/query/QueryRunner.ts
@@ -7,7 +7,7 @@ import { QueryOptions } from "mysql2/promise";
 import { QueryParser } from "./QueryParser";
 import { QuerySchema } from "../../../types/query.schema";
 import { TiDBQueryExecutor } from "../../executor/query-executor/TiDBQueryExecutor";
-import {dataQueryTimer, measure} from "../../../plugins/metrics/metrics";
+import {dataQueryTimer, measure} from "../../../plugins/metrics";
 
 export const enum QueryType {
   QUERY = 'query',
