--- packages/api-server/src/plugins/README.md
@@ -1,16 +1,0 @@
-# Plugins Folder
-
-Plugins define behavior that is common to all the routes in your
-application. Authentication, caching, templates, and all the other cross
-cutting concerns should be handled by plugins placed in this folder.
-
-Files in this folder are typically defined through the
-[`fastify-plugin`](https://github.com/fastify/fastify-plugin) module,
-making them non-encapsulated. They can define decorators and set hooks
-that will then be used in the rest of your application.
-
-Check out:
-
-* [The hitchhiker's guide to plugins](https://www.fastify.io/docs/latest/Guides/Plugins-Guide/)
-* [Fastify decorators](https://www.fastify.io/docs/latest/Reference/Decorators/).
-* [Fastify lifecycle](https://www.fastify.io/docs/latest/Reference/Lifecycle/).
