+++ packages/api-server/src/plugins/redis.ts
@@ -4,7 +4,7 @@ import fastifyRedis from "@fastify/redis";
 export default fp(async (app) => {
     app.register(fastifyRedis, { url: app.config.REDIS_URL });
 }, {
-    name: '@fastify/redis',
+    name: 'fastify-redis',
     dependencies: [
         '@fastify/env'
     ]
